fix build.sql and add split function for static data
