+++ server/database/config/build.js
@@ -1,17 +1,20 @@
-const { readFileSync } = require("fs");
-const { join } = require("path");
+const { readFileSync } = require('fs');
+const { join } = require('path');
 
-const dbConnection = require("./db_connection");
+const dbConnection = require('./connection');
 
 const buildDb = () => {
-  const buildSql = readFileSync(join(__dirname, "build.sql")).toString();
-  const staticSql = readFileSync(join(__dirname, "staticData.sql")).toString();
-  return dbConnection.query(buildSql).then(() => dbConnection.query(staticSql));
+  const buildSql = readFileSync(join(__dirname, 'build.sql')).toString();
+  return dbConnection.query(buildSql);
 };
 
+const buildStaticData = () => {
+  const staticSql = readFileSync(join(__dirname, 'staticData.sql')).toString();
+  return dbConnection.query(staticSql);
+};
 const buildFakeData = () => {
-  const sql = readFileSync(join(__dirname, "fakeData.test.sql")).toString();
-  return dbConnection.query(sql);
+  const fakeSql = readFileSync(join(__dirname, 'fakeData.test.sql')).toString();
+  return dbConnection.query(fakeSql);
 };
 
-module.exports = { buildDb, buildFakeData };
+module.exports = { buildDb, buildFakeData, buildStaticData };
